Create batch_log_file_renamer.py
This is a utility for automatically decoding the 8 character filenames that are created by the datalogger. This script automatically renames all files in a directory with the appropriate extended format that tells the date, time, and recording mode (e.g., "H08L403B.csv" becomes "July 08, 2022 at 11-40-30 - Burst.csv"). Copy log files from the logger's SD card into a new directory for long term storage, then run this script, feeding it the path of the directory that contains the copied files.